# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvestigateRequest": {
        "properties": {
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "InvestigateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Agentic Cybersecurity Threat Intelligence & Incident Response Advisor",
    "title": "ACTIRA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/incidents/{incident_id}/investigate": {
      "post": {
        "operationId": "investigate_incident_api_incidents__incident_id__investigate_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Investigate Incident",
        "tags": [
          "investigate"
        ]
      }
    },
    "/api/incidents/{incident_id}/investigate/stream": {
      "post": {
        "description": "SSE stream of AI Investigator tokens + final structured answer.\n\nEvents (data JSON):\n  status | meta | token | done | error\n`done` includes answer (sanitized object) and persists the investigation.\nNon-streaming POST /investigate remains available as a fallback.",
        "operationId": "investigate_incident_stream_api_incidents__incident_id__investigate_stream_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Investigate Incident Stream",
        "tags": [
          "investigate"
        ]
      }
    },
    "/api/incidents/{incident_id}/investigations": {
      "get": {
        "operationId": "list_investigations_api_incidents__incident_id__investigations_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Investigations",
        "tags": [
          "investigate"
        ]
      }
    },
...
    "/api/investigate/starter-questions": {
      "get": {
        "operationId": "starter_questions_api_investigate_starter_questions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Starter Questions",
        "tags": [
          "investigate"
        ]
      }
    },
...
    "/api/logs/jobs/{job_id}/events": {
      "get": {
        "description": "SSE stream of job phase updates (not LLM tokens).\n\nPolls Mongo until status is terminal (done / failed / error) or timeout.\nUseful for upload UI progress without tight client polling loops.",
        "operationId": "job_phase_events_api_logs_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Job Phase Events",
        "tags": [
          "investigate"
        ]
      }
    },
...
    "/api/v1/incidents/{incident_id}/investigate": {
      "post": {
        "operationId": "investigate_incident_api_v1_incidents__incident_id__investigate_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Investigate Incident",
        "tags": [
          "investigate"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/investigate/stream": {
      "post": {
        "description": "SSE stream of AI Investigator tokens + final structured answer.\n\nEvents (data JSON):\n  status | meta | token | done | error\n`done` includes answer (sanitized object) and persists the investigation.\nNon-streaming POST /investigate remains available as a fallback.",
        "operationId": "investigate_incident_stream_api_v1_incidents__incident_id__investigate_stream_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Investigate Incident Stream",
        "tags": [
          "investigate"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/investigations": {
      "get": {
        "operationId": "list_investigations_api_v1_incidents__incident_id__investigations_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Investigations",
        "tags": [
          "investigate"
        ]
      }
    },
...
    "/api/v1/investigate/starter-questions": {
      "get": {
        "operationId": "starter_questions_api_v1_investigate_starter_questions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Starter Questions",
        "tags": [
          "investigate"
        ]
      }
    },
...
    "/api/v1/logs/jobs/{job_id}/events": {
      "get": {
        "description": "SSE stream of job phase updates (not LLM tokens).\n\nPolls Mongo until status is terminal (done / failed / error) or timeout.\nUseful for upload UI progress without tight client polling loops.",
        "operationId": "job_phase_events_api_v1_logs_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Job Phase Events",
        "tags": [
          "investigate"
        ]
      }
    },
...
  }
}